recent update

+++ app/routers/post.py
@@ -1,5 +1,6 @@
 from fastapi import FastAPI, Response,status,HTTPException, Depends, APIRouter
 from .. import models, schemas, utils
+from typing import Optional
 
 from ..database import get_db
 from sqlalchemy.orm import Session
@@ -17,11 +18,11 @@ router = APIRouter(
 
 
 @router.get("/", response_model=list[schemas.Post])
-def get_posts(db: Session = Depends(get_db), current_user: int = Depends(utils.get_current_user)):
+def get_posts(db: Session = Depends(get_db), current_user: int = Depends(utils.get_current_user), limit: int = 10, skip: int = 0, search: Optional[str] = ""):
     # cursor.execute("SELECT * FROM posts")
     # posts = cursor.fetchall()
     # print(posts)
-    posts = db.query(models.Post).all()
+    posts = db.query(models.Post).filter(models.Post.title.contains(search)).limit(limit).offset(skip).all()
     return posts
 
 
@@ -43,8 +44,8 @@ def create_posts(new_post: schemas.PostCreate, db: Session = Depends(get_db), cu
     # post = cursor.fetchone()
     # conn.commit()
         # post = models.Post(title=new_post.title, content=new_post.content, published=new_post.published)
-        post = models.Post(**new_post.model_dump())
-        post.user_id = current_user
+        post = models.Post(user_id=current_user.id, **new_post.model_dump())
+        
         db.add(post)
         db.commit()
         db.refresh(post)
@@ -78,6 +79,9 @@ def delete_posts(id: int, db: Session = Depends(get_db), current_user: int = Dep
 
     if not deleted_post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id: {id} does not exist")
+    
+    if deleted_post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     # my_posts.pop(index)
     post_query.delete(synchronize_session=False)
     db.commit()
@@ -98,6 +102,9 @@ def update_posts(id: int, post: schemas.PostCreate, db: Session = Depends(get_db
     updated_post = query_post.first()
     if not updated_post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id: {id} does not exist")
+    
+    if updated_post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
 #    post_dict = post.model_dump()
 #    post_dict['id']=id
 #    my_posts[index]=post_dict
